pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motoko"
description = "A modular Discord bot"
version = "2025.8.29"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Fritz", email = "fritz@g01iath.online"}
]

dependencies = [
    "discord.py>=2.6.2",
    "requests>=2.32.5",
    "tzdata",
]

[project.urls]
Repository = "https://github.com/fritztafer/motoko"